+++ src/test/java/com/iexec/sms/tee/config/TeeWorkerInternalConfigurationTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2024 IEXEC BLOCKCHAIN TECH
+ * Copyright 2022-2025 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,7 +20,11 @@ import com.iexec.commons.poco.tee.TeeFramework;
 import com.iexec.sms.api.config.GramineServicesProperties;
 import com.iexec.sms.api.config.SconeServicesProperties;
 import com.iexec.sms.api.config.TeeAppProperties;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.springframework.util.unit.DataSize;
+
+import java.util.Collections;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -29,7 +33,7 @@ class TeeWorkerInternalConfigurationTests {
     private static final String FINGERPRINT = "fingerprint";
     private static final String ENTRYPOINT = "entrypoint";
     private static final String LAS_IMAGE = "lasImage";
-    private static final long HEAP_SIZE_GB = 3;
+    private static final DataSize HEAP_SIZE = DataSize.parse("3GB");
     private static final long HEAP_SIZE_B = 3221225472L;
 
     private final TeeAppProperties preComputeProperties = TeeAppProperties.builder()
@@ -45,36 +49,46 @@ class TeeWorkerInternalConfigurationTests {
             .heapSizeInBytes(1L)
             .build();
 
+    private TeeWorkerInternalConfiguration teeWorkerInternalConfiguration;
+    private TeeWorkerPipelineConfiguration pipelineConfig;
 
-    private final TeeWorkerInternalConfiguration teeWorkerInternalConfiguration =
-            new TeeWorkerInternalConfiguration();
+    @BeforeEach
+    void setUp() {
+        teeWorkerInternalConfiguration = new TeeWorkerInternalConfiguration();
 
-    // region preComputeProperties
-    @Test
-    void preComputeProperties() {
-        TeeAppProperties properties = teeWorkerInternalConfiguration.preComputeProperties(
+        final TeeWorkerPipelineConfiguration.StageConfig validStageConfig = new TeeWorkerPipelineConfiguration.StageConfig(
                 IMAGE,
                 FINGERPRINT,
-                ENTRYPOINT,
-                HEAP_SIZE_GB
+                HEAP_SIZE,
+                ENTRYPOINT
         );
+
+        final TeeWorkerPipelineConfiguration.Pipeline pipeline = new TeeWorkerPipelineConfiguration.Pipeline(
+                "v5",
+                validStageConfig,
+                validStageConfig
+        );
+
+        pipelineConfig = new TeeWorkerPipelineConfiguration(
+                Collections.singletonList(pipeline)
+        );
+    }
+
+    // region teeWorkerInternalConfiguration
+
+    @Test
+    void shouldGetPreComputePropertiesFromPipeline() {
+        final TeeAppProperties properties = teeWorkerInternalConfiguration.preComputeProperties(pipelineConfig);
 
         assertEquals(IMAGE, properties.getImage());
         assertEquals(FINGERPRINT, properties.getFingerprint());
         assertEquals(ENTRYPOINT, properties.getEntrypoint());
         assertEquals(HEAP_SIZE_B, properties.getHeapSizeInBytes());
     }
-    // endregion
 
-    // region postComputeProperties
     @Test
-    void postComputeProperties() {
-        TeeAppProperties properties = teeWorkerInternalConfiguration.postComputeProperties(
-                IMAGE,
-                FINGERPRINT,
-                ENTRYPOINT,
-                HEAP_SIZE_GB
-        );
+    void shouldGetPostComputePropertiesFromPipeline() {
+        final TeeAppProperties properties = teeWorkerInternalConfiguration.postComputeProperties(pipelineConfig);
 
         assertEquals(IMAGE, properties.getImage());
         assertEquals(FINGERPRINT, properties.getFingerprint());
@@ -85,8 +99,8 @@ class TeeWorkerInternalConfigurationTests {
 
     // region gramineServicesProperties
     @Test
-    void gramineServicesProperties() {
-        GramineServicesProperties properties =
+    void shouldBuildGramineServicesProperties() {
+        final GramineServicesProperties properties =
                 teeWorkerInternalConfiguration.gramineServicesProperties(preComputeProperties, postComputeProperties);
 
         assertEquals(TeeFramework.GRAMINE, properties.getTeeFramework());
@@ -97,8 +111,8 @@ class TeeWorkerInternalConfigurationTests {
 
     // region sconeServicesProperties
     @Test
-    void sconeServicesProperties() {
-        SconeServicesProperties properties =
+    void shouldBuildSconeServicesProperties() {
+        final SconeServicesProperties properties =
                 teeWorkerInternalConfiguration.sconeServicesProperties(preComputeProperties, postComputeProperties, LAS_IMAGE);
 
         assertEquals(TeeFramework.SCONE, properties.getTeeFramework());